requirements.txt
moviepy==1.0.3
imageio[ffmpeg]>=2.31
numpy>=1.24
requests>=2.31
pydub>=0.25
Pillow>=10.0
httpx>=0.27
